+++ drivers/net/txgbe/txgbe_ethdev.c
@@ -3764,7 +3764,7 @@ txgbe_configure_msix(struct rte_eth_dev *dev)
 
 int
 txgbe_set_queue_rate_limit(struct rte_eth_dev *dev,
-			   uint16_t queue_idx, uint16_t tx_rate)
+			   uint16_t queue_idx, uint32_t tx_rate)
 {
 	struct txgbe_hw *hw = TXGBE_DEV_HW(dev);
 	uint32_t bcnrc_val;
